Add files via upload
The project implements a restaurant management module focusing on the stock of raw materials in the kitchen. Customers have the option to consult a menu containing all the dishes they can order. Each dish has a recipe consisting of the raw materials and the quantities needed for preparation. A customer can add multiple items from the menu to an order. The program checks the availability of raw materials and processes the orders in the system accordingly. The entire flow is managed based on console commands defined by me in a legend.